Fixes to prediction (#16)
## Purpose 

To improve the predictions, we slightly modified the handling of grasses. Of notables importance is the estimation of the absolute minimum temperature from monthly temperatures rather than temperature of the coldest month, which greatly affects biome ranges. We also shifted how we save the PFTvalues, as the indexing was shifted by one and not properly saved into the NetCDF file. 

## Code changes:

- 'biome4driver': no more tminfile as input, and new indexing 
- 'competition2': adapted the fill of the output file to the indexing
- 'plotting': increasing the contrast between colors
- 'other files': matching more closely the original BIOME4 code

## Review

For the review process follow the guidelines at [Checklist](https://meteoswiss.atlassian.net/wiki/spaces/UA/pages/20156488/Code+Review)