Changes to dominance and BasePFTs (#7)
Purpose
Following a discussion with Nick and Philipp on 22/07/2025, we're shifting the Base PFTs from climate-based ones to physiology-based ones.

We're also adding a 20th characteristic to the PFTs, which is the dominance factor. This is based on the work of Haxeltine and Prentice (1992) on the first Biome model. They rank the PFTs based on a factor d, which we reintroduce here. This relates to the PFT's capacity to dominate even when it is at the edge of its range. Some species, like the birch, can dominate even under stress, while others are not, and this should be reflected in the competitive interactions.