Interface for Package (#6)
# Purpose 
We are creating an interface to run the model once we have loaded the package to make it as straightforward for the user to set up their PFTList, choose their model and load whatever file type their want as a raster. 

# Code Changes
The driver, which originally was a script is now `Biomedriver.jl`, which is a module that calls Biome and provides an interface.