Matching current BIOME4 state (#10)
# Purpose
Adapting the assign biome logic to the current state of BIOME4 and handling tmin as an input instead of always calculating it from temp. 

## Note 

There are still issues with the handling of strx and stry and the cut of the input data that will need to be resolved as soon as possible.